pyproject.toml
[project]
name = "data-engineering-agents"
version = "0.1.0"
description = "Autonomous Data Engineering Agents — Medallion Architecture on GCP"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-bigquery>=3.25,<4",
    "google-cloud-storage>=2.18,<3",
    "google-cloud-pubsub>=2.23,<3",
    "google-cloud-firestore>=2.19,<3",
    "google-cloud-datacatalog-lineage>=0.3,<1",
    "google-cloud-logging>=3.11,<4",
    "google-cloud-aiplatform>=1.72,<2",
    "google-cloud-dataflow-client>=0.8,<1",
    "pydantic>=2.9,<3",
    "pydantic-settings>=2.6,<3",
    "pyyaml>=6.0,<7",
    "jinja2>=3.1,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<1",
]

[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
